fix: fixing many functions to be comply with the norm
Fixed up new mini split. Splitting up its biggger functions into smaller ones. Handling certain errors more correctly as well. Also began work and I believe I am even finished with fixing up the split cleaner. Leaving only a few other things left to finish up and clean up.
