+++ second_try/minishell.h
@@ -6,7 +6,7 @@
 /*   By: akuburas <akuburas@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 11:16:05 by akuburas          #+#    #+#             */
-/*   Updated: 2024/04/30 13:44:16 by akuburas         ###   ########.fr       */
+/*   Updated: 2024/05/02 06:20:16 by akuburas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -180,6 +180,7 @@ typedef struct s_new_string_data
 	char			*new_string;
 	t_input_list	*temp;
 	t_env_list		*env;
+	t_shelldata		*shell_data;
 }	t_new_string_data;
 
 //A struct for the mini_split
@@ -227,8 +228,13 @@ int			split_cleaner(t_shelldata *data);
 void		copy_dollar(t_new_string_data *data);
 t_env_list	*try_to_find_env(t_env_list *env, char *str, int len);
 void		potential_split_create(t_new_string_data *data);
+void		split_env(t_new_string_data *data, t_env_list	*temp_env);
 
+
+
+//new mini split functions
 int			new_mini_split(t_shelldata *data);
+int			duplicate_input(char *input, t_shelldata *data, int *i);
 
 
 //signal handler
@@ -277,5 +283,10 @@ int			create_2d_env(t_shelldata *data);
 int			ft_echo(t_child_data *data, int fd);
 int			child_pre_check(t_shelldata *data);
 int			ft_pwd(char *pwd);
+
+//error functions
+int			check_pipes(t_shelldata *data);
+void		split_memory_failed(t_shelldata *data);
+
 
 #endif
