+++ second_try/split_cleaner.c
@@ -6,7 +6,7 @@
 /*   By: akuburas <akuburas@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 09:56:31 by akuburas          #+#    #+#             */
-/*   Updated: 2024/04/30 13:49:34 by akuburas         ###   ########.fr       */
+/*   Updated: 2024/05/02 05:57:29 by akuburas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,15 +25,8 @@ void	new_str_quote_found(t_new_string_data *data, char *temp)
 	}
 }
 
-int	new_string_last_check(t_input_list *temp)
-{
-	if (temp->word_split == FAILURE)
-		return (NO_MEMORY);
-	return (SUCCESS);
-}
-
 void	set_up_string_data(t_new_string_data *data, t_input_list *temp,
-	t_env_list *env, char *new_string)
+	t_shelldata *shell_data, char *new_string)
 {
 	data->i = 0;
 	data->j = 0;
@@ -41,6 +34,7 @@ void	set_up_string_data(t_new_string_data *data, t_input_list *temp,
 	data->temp = temp;
 	data->env = env;
 	data->new_string = new_string;
+	data->shell_data = shell_data;
 }
 
 void	handle_dollar_sign(t_new_string_data *data)
@@ -51,11 +45,11 @@ void	handle_dollar_sign(t_new_string_data *data)
 		copy_dollar(data);
 }
 
-int	create_new_string(t_input_list *temp, t_env_list *env, char *n_str)
+int	create_new_string(t_input_list *temp, t_shelldata *shell_data, char *n_str)
 {
 	t_new_string_data	data;
 
-	set_up_string_data(&data, temp, env, n_str);
+	set_up_string_data(&data, temp, shell_data, n_str);
 	while (data.temp->input[data.j] != '\0')
 	{
 		if (ft_strchr("'\"", data.temp->input[data.j]) != NULL)
@@ -67,16 +61,26 @@ int	create_new_string(t_input_list *temp, t_env_list *env, char *n_str)
 		}
 		else
 			handle_dollar_sign(&data);
-		if (new_string_last_check(data.temp) == NO_MEMORY)
-			return (NO_MEMORY);
 		data.j++;
 	}
-	if (data.temp->word_split != WORD_SPLIT)
-		data.temp->old_input = temp->input;
-	else
-		free(data.temp->input);
+	data.temp->old_input = temp->input;
 	data.temp->input = data.new_string;
 	return (SUCCESS);
+}
+
+void	clean_up_split(t_shelldata *data, t_input_list *temp)
+{
+	size_t	length;
+	char	*str;
+
+	str = (char *)ft_calloc(length + 1, sizeof(char));
+	if (str == NULL)
+		split_memory_failed(data);
+	if (create_new_string(temp, data, str) != SUCCESS)
+	{
+		clear_input(data->input_list, NO_MEMORY);
+		return (NO_MEMORY);
+	}
 }
 
 int	split_cleaner(t_shelldata *data)
@@ -90,19 +94,7 @@ int	split_cleaner(t_shelldata *data)
 	{
 		length = new_length(temp, data->env_list);
 		if (temp->needs_cleaning != 0)
-		{
-			str = (char *)ft_calloc(length + 1, sizeof(char));
-			if (str == NULL)
-			{
-				clear_input(data->input_list, NO_MEMORY);
-				return (NO_MEMORY);
-			}
-			if (create_new_string(temp, data->env_list, str) != SUCCESS)
-			{
-				clear_input(data->input_list, NO_MEMORY);
-				return (NO_MEMORY);
-			}
-		}
+			clean_up_split(data, temp);
 		temp = temp->next;
 		while (temp != NULL && temp->word_split == WORD_SPLIT)
 			temp = temp->next;
